# contracts/public.openapi.yaml
openapi: 3.1.0
info:
  title: redAI Personal owner API
  version: 1.0.0-draft.1
  description: Normative interface specification, not an implemented service. Domain
    authorization remains mandatory.
servers:
- url: /api/v1
security:
- OwnerSession: []
paths:
...
  /providers/{provider_id}/probe:
    post:
      operationId: probeProvider
      summary: probeProvider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./schemas/api.schema.json#/$defs/ProbeRequest
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: ./schemas/api.schema.json#/$defs/ProbeResult
        default:
          description: Structured error
          content:
            application/json:
              schema:
                $ref: ./schemas/common.schema.json#/$defs/Error
      parameters:
      - name: provider_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-CSRF-Token
        in: header
        required: true
        schema: *id001
      - name: Idempotency-Key
        in: header
        required: true
        schema: *id002
...
  /secrets/{secret_id}:
    delete:
      operationId: revokeSecret
      summary: revokeSecret
      responses:
        '204':
          description: No content
        default:
          description: Structured error
          content:
            application/json:
              schema:
                $ref: ./schemas/common.schema.json#/$defs/Error
      parameters:
      - name: secret_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-CSRF-Token
        in: header
        required: true
        schema: *id001
      - name: Idempotency-Key
        in: header
        required: true
        schema: *id002
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./schemas/api.schema.json#/$defs/Reason
...
  /settings:
    get:
      operationId: getSettings
      summary: getSettings
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: ./schemas/api.schema.json#/$defs/Settings
        default:
          description: Structured error
          content:
            application/json:
              schema:
                $ref: ./schemas/common.schema.json#/$defs/Error
    patch:
      operationId: updateSettings
      summary: updateSettings
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: ./schemas/api.schema.json#/$defs/Settings
        default:
          description: Structured error
          content:
            application/json:
              schema:
                $ref: ./schemas/common.schema.json#/$defs/Error
      parameters:
      - name: X-CSRF-Token
        in: header
        required: true
        schema: *id001
      - name: If-Match
        in: header
        required: true
        schema: *id003
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./schemas/api.schema.json#/$defs/SettingsUpdate
...
components:
  securitySchemes:
    OwnerSession:
      type: apiKey
      in: cookie
      name: __Host-redai_session